Fix big merge

+++ app/src/main/java/fr/vyfe/activity/EditSessionActivity.java
@@ -13,6 +13,8 @@ import android.text.TextWatcher;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.ImageView;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import com.google.android.gms.tasks.Continuation;
@@ -36,12 +38,14 @@ public class EditSessionActivity extends VyfeActivity {
         setContentView(R.layout.activity_info_video);
 
         Button btnCancel = findViewById(R.id.btn_cancel);
-        Button btnConfirmDelete = findViewById(R.id.btn_confirm_delete);
+        final Button btnConfirmDelete = findViewById(R.id.btn_confirm_delete);
         Button btnDelete = findViewById(R.id.btn_delete);
-        Button btnEdit = findViewById(R.id.bt_edit);
+        final Button btnEdit = findViewById(R.id.bt_edit);
         final ConstraintLayout confirmDelete = findViewById(R.id.confirm_delete);
         final EditText etDescription = findViewById(R.id.et_description);
         final EditText etSessionTitle = findViewById(R.id.et_video_title);
+        final TextView tvSave = findViewById(R.id.tv_video_save);
+        final ImageView imageViewDelete = findViewById(R.id.iv_delete);
         Toolbar toolbar = findViewById(R.id.toolbar);
         mDatabase = FirebaseDatabase.getInstance();
 
@@ -50,14 +54,51 @@ public class EditSessionActivity extends VyfeActivity {
 
         viewModel = ViewModelProviders.of(this, new EditSessionViewModelFactory(mAuth.getCurrentUser().getCompany())).get(EditSessionViewModel.class);
         viewModel.init(getIntent().getStringExtra(Constants.SESSIONMODELID_EXTRA));
-
         viewModel.getSession().observe(this, new Observer<SessionModel>() {
             @Override
             public void onChanged(@Nullable SessionModel sessionModel) {
-                if(sessionModel!=null){
+                if (sessionModel != null) {
                     etDescription.setText(sessionModel.getDescription());
                     etSessionTitle.setText(sessionModel.getName());
                 }
+
+                etDescription.addTextChangedListener(new TextWatcher() {
+                    @Override
+                    public void beforeTextChanged(CharSequence s, int start, int count, int after) {
+
+                    }
+
+                    @Override
+                    public void onTextChanged(CharSequence s, int start, int before, int count) {
+
+                    }
+
+                    @Override
+                    public void afterTextChanged(Editable s) {
+                        btnEdit.setClickable(true);
+                        btnEdit.setEnabled(true);
+                        btnEdit.setAlpha(1);
+                        viewModel.setNewDescription(s.toString());
+                    }
+                });
+
+                etSessionTitle.addTextChangedListener(new TextWatcher() {
+                    @Override
+                    public void beforeTextChanged(CharSequence s, int start, int count, int after) {
+
+                    }
+
+                    @Override
+                    public void onTextChanged(CharSequence s, int start, int before, int count) {
+
+                    }
+
+                    @Override
+                    public void afterTextChanged(Editable s) {
+                        btnEdit.setAlpha(1);
+                        viewModel.setNewName(s.toString());
+                    }
+                });
             }
         });
 
@@ -65,25 +106,43 @@ public class EditSessionActivity extends VyfeActivity {
             @Override
             public void onClick(View v) {
                 confirmDelete.setVisibility(View.VISIBLE);
+                imageViewDelete.setVisibility(View.VISIBLE);
             }
         });
 
         btnConfirmDelete.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                 viewModel.deleteSession().continueWith(new Continuation<Void, Void>() {
-                     @Override
-                     public Void then(@NonNull Task<Void> task) throws Exception {
-                         if (task.isSuccessful()) {
-                             Toast.makeText(EditSessionActivity.this, "Vidéo supprimée", Toast.LENGTH_SHORT).show();
-                             Intent intent = new Intent(EditSessionActivity.this, MySessionsActivity.class);
-                             startActivity(intent);
-                         }
-                         else
-                             Toast.makeText(EditSessionActivity.this, "Error", Toast.LENGTH_SHORT).show();
-                         return null;
-                     }
-                 });
+
+                if (btnConfirmDelete.getText().equals(R.string.delete)) {
+                    viewModel.deleteSession().continueWith(new Continuation<Void, Void>() {
+                        @Override
+                        public Void then(@NonNull Task<Void> task) throws Exception {
+                            if (task.isSuccessful()) {
+                                Toast.makeText(EditSessionActivity.this, R.string.delete_session, Toast.LENGTH_LONG).show();
+                                Intent intent = new Intent(EditSessionActivity.this, MySessionsActivity.class);
+                                startActivity(intent);
+                            } else
+                                Toast.makeText(EditSessionActivity.this, R.string.error, Toast.LENGTH_SHORT).show();
+                            return null;
+                        }
+                    });
+                } else {
+
+                    viewModel.editSession().continueWith(new Continuation<Void, Void>() {
+                        @Override
+                        public Void then(@NonNull Task<Void> task) throws Exception {
+                            if (task.isSuccessful()) {
+                                Toast.makeText(EditSessionActivity.this, R.string.save_edit, Toast.LENGTH_LONG).show();
+                                Intent intent = new Intent(EditSessionActivity.this, SelectVideoActivity.class);
+                                intent.putExtra(Constants.SESSIONMODELID_EXTRA, viewModel.getSession().getValue().getId());
+                                EditSessionActivity.this.startActivity(intent);
+                            } else
+                                Toast.makeText(EditSessionActivity.this, R.string.error, Toast.LENGTH_SHORT).show();
+                            return null;
+                        }
+                    });
+                }
             }
         });
 
@@ -91,59 +150,19 @@ public class EditSessionActivity extends VyfeActivity {
             @Override
             public void onClick(View v) {
                 confirmDelete.setVisibility(View.GONE);
-            }
-        });
-
-        etDescription.addTextChangedListener(new TextWatcher() {
-            @Override
-            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-
-            }
-
-            @Override
-            public void onTextChanged(CharSequence s, int start, int before, int count) {
-
-            }
-
-            @Override
-            public void afterTextChanged(Editable s) {
-                viewModel.setNewDescription(s.toString());
-            }
-        });
-
-        etSessionTitle.addTextChangedListener(new TextWatcher() {
-            @Override
-            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-
-            }
-
-            @Override
-            public void onTextChanged(CharSequence s, int start, int before, int count) {
-
-            }
-
-            @Override
-            public void afterTextChanged(Editable s) {
-                viewModel.setNewName(s.toString());
+                btnConfirmDelete.setText(R.string.delete);
+                tvSave.setText(R.string.confirm_delete_video);
             }
         });
 
         btnEdit.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                viewModel.editSession().continueWith(new Continuation<Void, Void>() {
-                    @Override
-                    public Void then(@NonNull Task<Void> task) throws Exception {
-                        if (task.isSuccessful())
-                            Toast.makeText(EditSessionActivity.this, R.string.save_session, Toast.LENGTH_SHORT).show();
-                        else
-                            Toast.makeText(EditSessionActivity.this, "Error", Toast.LENGTH_SHORT).show();
-                        return null;
-                    }
-                });
+                btnConfirmDelete.setText(R.string.change_movie);
+                tvSave.setText(R.string.confirm_change_movie);
+                confirmDelete.setVisibility(View.VISIBLE);
+                imageViewDelete.setVisibility(View.GONE);
             }
         });
-
     }
-
 }
